* src/include/nonposix.h (fileno) [_MSC_VER]: Removed.
* font/devlbp/Makefile.sub (CLEANADD): Set it.
* tmac/Makefile.sub (CLEANADD): Add tmac.local-s.
* Makefile.in (dist): Don't remove src/xditview/Imakefile, but ...
* Makefile.sub (DISTCLEANFILES): Here.

* libs/libgroff/new.cc: Include `nonposix.h'.
* win32-diffs: Updated.

When inside displays, an empty input line warning should be
suppressed.  If another macro call is put inside a display,
all subsequent empty lines found in that display caused a
warning to be emitted.

(doc-empty-line): Check the `doc-display-depth' register to
determine whether we are inside display or not.
the only change is that `.Fn' and `.Fc' put a final semicolon after
a function declaration in the SYNOPSIS section).