* release of groffer 1.0.0
Changements of the shell version since 0.9.31

* shell/groffer.sh: Use version.sh in the upper directory.


The groffer versions 1.* have two kinds of program, a shell
version and a Perl version.

The shell version is kept from the old shell only versions of
groffer 0.*.  Most of the former files in the main groffer
directory are now moved to the subdirectoy shell/.

The Perl version of groffer is a complete rewrite; most of its
files are found in the subdirectory perl/.

The Makefile.sub decides whether the shell or Perl version of
groffer is installed.  This is done by the program perl_test.pl.

In the following, all files in the groffer directory tree are
mentioned.

* ChangeLog: This file.  It contains information for groffer
versions >= 1.0.0 with shell and Perl kinds.  For older groffer
versions, see shell/ChangeLog.0 which contains information for the
shell only versions of groffer 0.*.

* Makefile.sub: The old Makefile.sub was extended to support the
shell and the Perl version of groffer at the same time.  If the
test of perl_test.pl succeeds the groffer Perl version will be
installed, otherwise the shell version is used.

* perl_test.pl: This is used by Makefile.sub and installed with
the Perl version of groffer.  It is a test of the installed perl
version.  The installed Perl version should be greater or equal
than the version that is required by this file.

* version.sh: This is the old file kept from the shell version of
groffer.  It stores the groffer version, the date of the last
update, and the groff version.  It is used and installed for the
shell and Perl version of groffer.

* README: This is the old README file extended by information on
the shell and Perl kinds of groffer.

* perl/: Subdirectory for the Perl version of groffer.

* perl/groffer.pl: This is the groffer script of the Perl
version, a Perl source file that handles the complete groffer
functionality.

* perl/man.pl: This is the collection of functions that are
related to man pages, apropos, and whatis.  It is loaded by
perl/groffer.pl.

* perl/func.pl: This is the collection of miscellaneous functions.
It is loaded by perl/groffer.pl.

* perl/split_env.sh: A shell script that is used by
perl/groffer.pl to split a large shell environment variable to a
Perl array.

* perl/groffer.man: This is the man page of the Perl version of
groffer.  It is derived from groffer.man of the shell version.  It
will only be installed when the Makefile.sub chooses to install
the Perl version of groffer instead of the shell version.

* perl/README_PERL: This file contains information of the Perl
compatibility and details that are special to the Perl version.

* shell/: Subdirectory for the shell version of groffer.

* shell/ChangeLog.0: The former ChangeLog file of the groffer
versions 0.* was moved to this file.  It contains information of
the shell only version of groffer 0.*.

* shell/groffer.sh: This is the old groffer.sh file of the shell
version.

* shell/groffer2.sh: This is the old groffer2.sh file of the shell
version.

* shell/groffer.man: This is the old man page groffer.man of the
shell version.  The shell version keeps its own man page.  It will
only be installed when the Makefile.sub chooses to install the
shell version instead of the Perl version.

* shell/README_SH: This is the old README_SH file of the shell
version containing information of the shell compatibility.