[pic] Fix a border case for arc computation.
* src/preproc/pic/object.cpp (object_spec::make_arc): It can happen
that test `radius < d' in the loop is satisfied, but the difference
in the two values is on the order of 1e-10.  If `radius' is small,
doubling the value can lead to a fairly gross error.

The original code appears to have been intended to deal with the
situation when radius is orders of magnitude less than `d'.  The
replacement code simply assigns `radius' the smallest value that
avoids problems with the floating point code further on in the
routine.