Handle ditroff command `Dt' without argument gracefully.
Without this patch, grotty goes into an infinite loop if it tries to
process

  x T utf8
  x res 240 24 40
  x init
  p1
  Dt

* src/libs/libdriver/input.cpp (get_integer_arg): Emit a fatal error
on a non-integer argument, bringing the code into line with the
behaviour documented in the header comment.
(get_possibly_integer_args): Terminate the loop on a non-integer
argument.
(next_arg_begin): Return newline or EOF after emitting the
corresponding error, rather than continuing on to the next line.