Fix Savannah bug #44768 (Solaris 10 compatibility issues).
* m4/groff.m4 (GROFF_DIFF_D): new macro to check if `diff' has
option -D, and search for `gdiff' if not (new variable DIFF_PROG
will contain the `diff' program to be used).

* m4/groff.m4 (GROFF_HAVE_TEST_EF_OPTION): new macro to check if
`test' supports the option -ef (created new variable
HAVE_TEST_EF_OPTION).

* m4/groff.m4 (GROFF_BASH): new macro to search for `bash' (new
variable BASH_PROG will contain the full path to `bash' or
`/bin/sh' if `bash' was not found).

* configure.ac: use the 3 new macros defined in m4/groff.m4.

* contrib/gdiffmk/gdiffmk.am: set HAVE_TEST_EF_OPTION, BASH_PROG,
DIFF_PROG in the generated script `gdiffmk'.

* contrib/gdiffmk/gdiffmk.sh:
- Fix incorrect space after '!' in sed script regex.
- `test -ef' is not portable: If `bash' is not available and the
shell does not support `test -ef', the option to pass a third
argument as the output file is disabled.
- Report an error if `diff' does not support -D option.
- GNU diff and Solaris diff support -D, but #ifndef / #endif are
treated differently (GNU adds a '!' in the #endif): modify the
regexs in the sed script accordingly.